tests/unit_tests/file/test_rbr.py
import pytest
import pandas as pd
import xarray as xr
from unittest.mock import MagicMock
from your_module import read  # Adjust the import as needed

@pytest.fixture
def mock_rsk_data(mocker):
    # Create a mock for the RSK class
    mock_rskdata = mocker.patch('pyrsktools.RSK').return_value
    mock_rskdata.open.return_value = None
    mock_rskdata.readdata.return_value = None
    mock_rskdata.data = {
        'timestamp': pd.to_datetime(['2023-01-01 00:00:00', '2023-01-02 00:00:00']),
        'conductivity': [10, 20],
        'temperature': [15, 16]
    }
    mock_rskdata.getchannelnamesandunits.return_value = (
        ['conductivity', 'temperature'],
        ['mS/cm', '°C']
    )
    mock_rskdata.instrument.model = 'RBR Model X'
    mock_rskdata.instrument.serialID = '123456789'
    mock_rskdata.scheduleInfo.samplingperiod.return_value = 60
    return mock_rskdata

def test_read(mock_rsk_data):
    # Call the read function
    ds_rsk = read('dummy_file.rsk')

    # Check if the Dataset has the correct variables and dimensions
    assert 'TIME' in ds_rsk.dims
    assert 'CNDC' in ds_rsk.variables
    assert 'TEMP' in ds_rsk.variables
    
    # Check the units and attributes of the variables
    assert ds_rsk['CNDC'].attrs['units'] == 'mS cm-1'
    assert ds_rsk['TEMP'].attrs['units'] == 'degC'
    assert ds_rsk['TIME'].attrs['units'] == 'days since 1970-01-01'
    assert ds_rsk['TIME'].attrs['axis'] == 'T'

    # Verify the attributes of the Dataset
    assert ds_rsk.attrs['instrument'] == 'RBR Model X'
    assert ds_rsk.attrs['instrument_serial_number'] == '123456789'
    assert ds_rsk.attrs['time_coverage_resolution'] == 'PT1M'  # Adjust if necessary

def test_empty_data(mocker):
    # Create a mock for the RSK class with empty data
    mock_rskdata = mocker.patch('pyrsktools.RSK').return_value
    mock_rskdata.open.return_value = None
    mock_rskdata.readdata.return_value = None
    mock_rskdata.data = {
        'timestamp': pd.to_datetime([]),
        'conductivity': [],
        'temperature': []
    }
    mock_rskdata.getchannelnamesandunits.return_value = (
        [], []
    )
    mock_rskdata.instrument.model = 'RBR Model X'
    mock_rskdata.instrument.serialID = '123456789'
    mock_rskdata.scheduleInfo.samplingperiod.return_value = 60

    # Call the read function
    ds_rsk = read('dummy_file.rsk')

    # Check if the Dataset is empty as expected
    assert ds_rsk.dims == {'TIME': 0}
    assert 'CNDC' not in ds_rsk.variables
    assert 'TEMP' not in ds_rsk.variables

def test_variable_renaming(mock_rsk_data):
    # Mock RBR_name_map to test renaming
    mock_rbr_name_map = {
        'conductivity': 'CNDC',
        'temperature': 'TEMP'
    }
    import your_module
    original_rbr_name_map = your_module.RBR_name_map
    your_module.RBR_name_map = mock_rbr_name_map

    try:
        ds_rsk = read('dummy_file.rsk')
        assert 'CNDC' in ds_rsk.variables
        assert 'TEMP' in ds_rsk.variables
    finally:
        your_module.RBR_name_map = original_rbr_name_map

def test_variable_units_mapping(mock_rsk_data):
    # Mock RBR_units_map to test unit conversion
    mock_rbr_units_map = {
        'mS/cm': 'mS cm-1',
        '°C': 'degC'
    }
    import your_module
    original_rbr_units_map = your_module.RBR_units_map
    your_module.RBR_units_map = mock_rbr_units_map

    try:
        ds_rsk = read('dummy_file.rsk')
        assert ds_rsk['CNDC'].attrs['units'] == 'mS cm-1'
        assert ds_rsk['TEMP'].attrs['units'] == 'degC'
    finally:
        your_module.RBR_units_map = original_rbr_units_map
